+++ internal/core/entity/ticket_view.go
@@ -5,6 +5,6 @@ import "time"
 type TicketView struct {
 	EventName    string
 	Token        string
-	TicketNumber int
+	TicketNumber int64
 	CheckedInAt  *time.Time
 }
