+++ internal/core/port/event_repository.go
@@ -6,10 +6,15 @@ import (
 )
 
 type EventRepository interface {
-	GetByID(id int) (entity.Event, error)
+	GetByID(id int64) (entity.Event, error)
+
 	GetByPublicID(publicID string) (entity.Event, error)
-	FindByIDForUpdate(tx *sql.Tx, id int) (entity.Event, error)
+
+	FindByIDForUpdate(tx *sql.Tx, id int64) (entity.Event, error)
+
 	FindByOwnerToken(token string) (entity.Event, error)
-	CountEventsByOrganizerEmail(email string) (int, error)
-	Create(event entity.Event) (int, error)
+
+	CountEventsByOrganizerEmail(email string) (int64, error)
+
+	Create(event entity.Event) (int64, error)
 }
